@@ -11,28 +11,30 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.0.5.RELEASE</version>
+        <version>3.1.1</version>
+        <relativePath/>
     </parent>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
-
-        <lombok.version>1.16.16</lombok.version>
+        <java.version>17</java.version>
+        <maven.compiler.source>${java.version}</maven.compiler.source>
+        <maven.compiler.target>${java.version}</maven.compiler.target>
+        <lombok.version>1.18.24</lombok.version>
         <guava.version>21.0</guava.version>
         <junit.version>4.12</junit.version>
         <typesafe.version>1.2.1</typesafe.version>
-        <postgresql.version>9.4-1206-jdbc42</postgresql.version>
+        <mysql.version>8.1.0</mysql.version>
         <mvel.version>2.4.4.Final</mvel.version>
     </properties>
 
     <dependencies>
-        <!--PostgreSql-->
+        <!--Mysql connector-->
         <dependency>
-            <groupId>org.postgresql</groupId>
-            <artifactId>postgresql</artifactId>
-            <version>${postgresql.version}</version>
+            <groupId>com.mysql</groupId>
+            <artifactId>mysql-connector-j</artifactId>
+            <version>${mysql.version}</version>
         </dependency>
 
         <!--Lombok Plugin-->
@@ -101,6 +103,7 @@
             <plugin>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-maven-plugin</artifactId>
+                <version>3.1.1</version>
             </plugin>
         </plugins>
     </build>
